+++ CHANGELOG.md
@@ -24,6 +24,7 @@
 - Refactor and simplify stream handover functionality to be more robust ([#3898](https://github.com/hoprnet/hoprnet/pull/3898))
 - Added possibility to specify custom RPC provider in Avado
 - Add connectivity health indicator & NR eligibility status of the node to the `info` command ([#3921](https://github.com/hoprnet/hoprnet/pull/3921))
+- Fix message encoding/decoding in HOPRd ([#3943](https://github.com/hoprnet/hoprnet/pull/3943))
 
 # Breaking changes